Fix templates not loading properly
